Update osh to 20081213, by maintainer's update request via PR 40157.
Upstream changes between osh-20081122 and osh-20081213 follow.

osh.c:
This release changes the effect of the `-v' command-line option so that,
in addition to printing non-blank command lines to the standard error,
it also prints each blank/empty command line as an empty line.

fd2.c, util.c:
This release changes the fd2 `-e' command-line option so that it
reflects the documented behavior.  For example, this change causes
invocation of `fd2 -ef file command' to redirect both conventional
output and diagnostic output from "command" to "file".